refactor(bzm2): extract hash thread bring-up flow
